fix(ci): restore validation worktree line limit
Split cleanup edge tests out of the oversized validation worktree test module so the maintainability shard passes without weakening the guard.
